+++ docs/api/changelog.rst
@@ -175,6 +175,9 @@ Version 0.11
 
 * Support for :xep:`27` schema in :mod:`aioxmpp.misc`
 
+* *Possible breaking change*: Renamed :meth:`aioxmpp.xso.XSO.unparse_to_sax` to
+  :meth:`~aioxmpp.xso.XSO.xso_serialise_to_sax`.
+
 .. _api-changelog-0.10:
 
 Version 0.10